Disable noice signature help. Too many issues..
The issues:
* cursor stealing - while typing in a signature, if I press ',' and then
' ', the cursor would jump into the signature help window no matter how
I asked noice to adjust settings. This may be solvable, but for now I'm
just going to disable it.

* weird window freakouts and leftovers - sometimes a signature help
window will not be deleted and just linger covering my text buffer. I'll
have to manually navigate to it to remove it.

* Sometimes if a signature help window has a scrollbar on it, that
scrollbar will not be deleted and just linger on the screen
indefinitely.

Bottom line: I need to figure out a better solution for nice looking
signature help...